Remove unnecessary type casts

+++ ex-app/bufferObject.c
@@ -15,10 +15,8 @@ stc_bffr_t0* const p_stc_bffr_g_buf0 = &stc_bffr_pv_buf0;
 stc_bffr_t1* const p_stc_bffr_g_buf1 = &stc_bffr_pv_buf1;
 
 /* Private storages for buffer objects */
-static cirquStrgElem_t0 a_cirquStrgElem_pv_buf0Strg[BUF0STRGSIZE] =
-    { (uint8_t)0U };
-static cirquStrgElem_t1 a_cirquStrgElem_pv_buf1Strg[BUF1STRGSIZE] =
-    { (uint32_t)0U };
+static cirquStrgElem_t0 a_cirquStrgElem_pv_buf0Strg[BUF0STRGSIZE] = {0U};
+static cirquStrgElem_t1 a_cirquStrgElem_pv_buf1Strg[BUF1STRGSIZE] = {0U};
 
 /* OPERATIONS
  * ==========
@@ -26,16 +24,14 @@ static cirquStrgElem_t1 a_cirquStrgElem_pv_buf1Strg[BUF1STRGSIZE] =
 
 void fn_bffrObj_buf0Ctor(void)
 {
-    fn_bffr_ini0(&stc_bffr_pv_buf0, a_cirquStrgElem_pv_buf0Strg,
-                (uint8_t)BUF0STRGSIZE);
+    fn_bffr_ini0(&stc_bffr_pv_buf0, a_cirquStrgElem_pv_buf0Strg, BUF0STRGSIZE);
 
     return;
 }
 
 void fn_bffrObj_buf1Ctor(void)
 {
-    fn_bffr_ini1(&stc_bffr_pv_buf1, a_cirquStrgElem_pv_buf1Strg,
-                (uint32_t)BUF1STRGSIZE);
+    fn_bffr_ini1(&stc_bffr_pv_buf1, a_cirquStrgElem_pv_buf1Strg, BUF1STRGSIZE);
 
     return;
 }
